requirements.txt
# Core runtime dependencies for reproducing the published experiments.
# If you need a CUDA-specific PyTorch build, install torch/torchvision
# from the appropriate PyTorch wheel index for your system.

accelerate==1.1.0
matplotlib==3.9.2
numpy==2.0.2
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
PyYAML==6.0.2
scikit-learn==1.5.2
seaborn==0.13.2
timm==1.0.19
torch==2.5.1
torcheval==0.0.7
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.66.6
transformers==4.46.2
umap-learn==0.5.7
wandb==0.18.5
